requirements.txt.txt
fastapi
uvicorn
requests
beautifulsoup4
psycopg2-binary